TestExplosion1

+++ Assets/Scripts/Controller.cs
@@ -4,7 +4,7 @@ using System.Collections;
 public class Controller: MonoBehaviour
 {
 			
-	public void Controlling (int speed, float move_x, float move_y)
+	public void Controlling (int speed, int move_x, int move_y)
 	{
 		rigidbody2D.velocity = new Vector2 (move_x * speed, move_y * speed);
